Restore R epiweekly aggregation path
