+++ tests/test_model_pipelines.py
@@ -29,14 +29,18 @@ def _common_kwargs(tmp_path):
     }
 
 
-def test_fable_pipeline_selects_weekly_nssp_inputs(tmp_path):
+@patch("cfa.stf.routine.fable.forecast_fable.generate_epiweekly_data")
+def test_fable_pipeline_aggregates_weekly_inputs(mock_generate, tmp_path):
     pipeline = FablePipeline(
         **_common_kwargs(tmp_path),
         n_samples=10,
         epiweekly=True,
     )
+    run = _run(tmp_path, model_name=pipeline.model_name)
 
-    assert pipeline.nssp_frequency == "epiweekly"
+    pipeline.transform_serialized_data(run)
+
+    mock_generate.assert_called_once_with(run.data_dir, overwrite_daily=True)
 
 
 @patch("cfa.stf.routine.fable.forecast_fable.fable_e_other_forecasts")
